Refactored the function for handleLogout as removeKeys in logout-helper.js
