+++ src/components/NavBar/NavBar.jsx
@@ -17,7 +17,7 @@ import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 import { useTheme } from '@mui/material/styles';
 // Utility Imports
-import { SOLID_IDENTITY_PROVIDER } from '../../utils';
+import { SOLID_IDENTITY_PROVIDER, removeKeys } from '../../utils';
 // Custom Hook Imports
 import { useRedirectUrl } from '../../hooks';
 // Component Imports
@@ -61,17 +61,7 @@ const NavBar = () => {
 
   // Event handler for logging out of SOLID POD and removing items from localStorage
   const handleLogout = () => {
-    const keysToRemove = [
-      'loggedIn',
-      'redirectUrl',
-      'restorePath',
-      `issuerConfig:${SOLID_IDENTITY_PROVIDER}`,
-      `issuerConfig:${SOLID_IDENTITY_PROVIDER.slice(0, -1)}`
-    ];
-
-    keysToRemove.forEach((key) => {
-      localStorage.removeItem(key);
-    });
+    removeKeys();
     setShowConfirmation(false);
   };
