+++ src/components/Notification/InactivityMessage.jsx
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 // Material UI Base imports; TODO: Update imports to @mui/material when re-styling
 import Button from '@mui/base/Button';
 // Utility imports
-import { SOLID_IDENTITY_PROVIDER } from '../../utils';
+import { removeKeys } from '../../utils';
 
 /**
  * Inactivity Notification Component - Component that displays a popup modal
@@ -63,17 +63,7 @@ const InactivityMessage = () => {
   // Event handler for logout and removing items from localStorage
   // Returns user to home page upon successful logout
   const handleLogout = () => {
-    const keysToRemove = [
-      'loggedIn',
-      'redirectUrl',
-      'restorePath',
-      `issuerConfig:${SOLID_IDENTITY_PROVIDER}`,
-      `issuerConfig:${SOLID_IDENTITY_PROVIDER.slice(0, -1)}`
-    ];
-
-    keysToRemove.forEach((key) => {
-      localStorage.removeItem(key);
-    });
+    removeKeys();
   };
 
   return (
