+++ src/utils/index.js
@@ -31,6 +31,7 @@ import {
 } from './session-core';
 import { SOLID_IDENTITY_PROVIDER } from './session-helper';
 import makeHandleFormSubmission from './FormSubmissionHelper';
+import removeKeys from './logout-helper';
 
 export {
   docTypes,
@@ -53,5 +54,6 @@ export {
   setDocContainerAclPermission,
   checkContainerPermission,
   sendMessageTTL,
-  makeHandleFormSubmission
+  makeHandleFormSubmission,
+  removeKeys
 };
